Update README.md
copy edits and additions.

I want to do a bit more on the use case descriptions before sharing more widely. In fact, I suggest we narrow this a bit initially to just supporting the Patient Day of Action, with the intent to generalize out after deeply understanding that use case and kicking ass this summer.